refactor(google): Move GoogleExecutor to google package
GoogleExecutor is only used by the GCE provider and has GCE-specific
logic; move it out of the common module so we can keep that module
containing only logic shared between google providers.
